agents/requirements-base.txt
# Pinned base requirements for all AAP agent containers
# Update versions here to propagate to all agents

# Microsoft Agent Framework (RC — pin exactly)
agent-framework==1.0.0rc5
agent-framework-orchestrations>=1.0.0b260311

# Azure AI Foundry SDK (GA)
# azure-ai-projects 2.x removed thread/run/message operations — use azure-ai-agents
azure-ai-projects>=2.0.1
azure-ai-agents>=1.0.0

# Foundry Hosted Agent adapters
# NOTE: Cannot pin to ==1.0.0b17 — that version declares agent-framework-core<=1.0.0rc3
# which conflicts with agent-framework==1.0.0rc5 (requires agent-framework-core==1.0.0rc5).
# Kept unpinned until Microsoft releases a b18+ build that supports rc4+/rc5.
# Monitor https://pypi.org/project/azure-ai-agentserver-agentframework/ and pin once
# a version compatible with agent-framework==1.0.0rc5 is available.
azure-ai-agentserver-core
azure-ai-agentserver-agentframework

# Azure Identity (managed identity via IMDS)
azure-identity>=1.17.0

# Cosmos DB SDK (session budget tracking)
azure-cosmos>=4.7.0

# OpenTelemetry + Azure Monitor
azure-monitor-opentelemetry>=1.6.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0

# MCP SDK (tool surface)
mcp[cli]>=1.26.0

# Data validation
pydantic>=2.8.0
